+++ spec/binding_spec.rb
@@ -28,10 +28,14 @@ describe "BindingTest" do
     # it should have its own interface
     expect(test2["org.ruby.Test2"]).not_to be_nil
     # but not an interface of the Test class
-    expect(test2["org.ruby.SampleInterface"]).to be_nil
+    expect { test2["org.ruby.SampleInterface"] }.to raise_error(DBus::Error) do |e|
+      expect(e.message).to match(/no such interface/)
+    end
 
     # and the parent should not get polluted by the child
-    expect(@base["org.ruby.Test2"]).to be_nil
+    expect { @base["org.ruby.Test2"] }.to raise_error(DBus::Error) do |e|
+      expect(e.message).to match(/no such interface/)
+    end
   end
 
   it "tests translating errors into exceptions" do